requirements.txt
fastapi
uvicorn
pydantic
google-auth
google-api-python-client
requests
